Move context files to src
Moved all files with concrete PaymentContext implementations from
context to src.

Created a new `Payment` namespace inside the `Infrastructure`namespace
to un-clutter that namespace a bit.

Deleted all mentions of the context directory from CI.
