+++ .travis.yml
@@ -30,7 +30,7 @@ script:
   - make setup
   - make ci
   - make install-php COMPOSER_FLAGS="--no-dev -q" # Remove dev dependencies to make sure PHPStan creates errors if prod code depends on dev classes
-  - docker run -v $PWD:/app --rm wmde/fundraising-frontend-phpstan analyse -c phpstan.neon --level 1 --no-progress cli/ contexts/ src/ # Can't use "make stan" because stan was removed
+  - docker run -v $PWD:/app --rm wmde/fundraising-frontend-phpstan analyse --level 1 --no-progress cli/ src/ # Can't use "make stan" because stan was removed
 
 after_success:
   - if [ "$TYPE" == "coverage" ]; then bash build/travis/uploadCoverage.sh; fi
