+++ composer.json
@@ -114,14 +114,12 @@
 		"psr-4": {
 			"WMDE\\Fundraising\\Frontend\\": "src/",
 			"WMDE\\Fundraising\\Frontend\\App\\": "app/",
-			"WMDE\\Fundraising\\Frontend\\Cli\\": "cli/",
-			"WMDE\\Fundraising\\PaymentContext\\": "contexts/PaymentContext/src/"
+			"WMDE\\Fundraising\\Frontend\\Cli\\": "cli/"
 		}
 	},
 	"autoload-dev": {
 		"psr-4": {
 			"WMDE\\Fundraising\\Frontend\\Tests\\": "tests/",
-			"WMDE\\Fundraising\\PaymentContext\\Tests\\": "contexts/PaymentContext/tests/",
 			"WMDE\\Fundraising\\SubscriptionContext\\Tests\\": "vendor/wmde/fundraising-subscriptions/tests/"
 		}
 	},
