+++ phpunit.xml.dist
@@ -14,11 +14,9 @@
     <testsuites>
         <testsuite name="unit">
             <directory>tests/Unit</directory>
-            <directory>contexts/*/tests/Unit</directory>
         </testsuite>
         <testsuite name="integration">
             <directory>tests/Integration</directory>
-            <directory>contexts/*/tests/Integration</directory>
         </testsuite>
         <testsuite name="edgetoedge">
             <directory>tests/EdgeToEdge</directory>
