+++ src/Infrastructure/Payment/KontoCheckIbanValidator.php
@@ -2,8 +2,9 @@
 
 declare( strict_types = 1 );
 
-namespace WMDE\Fundraising\PaymentContext\Domain;
+namespace WMDE\Fundraising\Frontend\Infrastructure\Payment;
 
+use WMDE\Fundraising\PaymentContext\Domain\IbanValidator;
 use WMDE\Fundraising\PaymentContext\Domain\Model\Iban;
 use WMDE\FunValidators\ConstraintViolation;
 use WMDE\FunValidators\ValidationResult;
