+++ src/Infrastructure/Payment/PayPalPaymentNotificationVerifierException.php
@@ -2,7 +2,7 @@
 
 declare( strict_types = 1 );
 
-namespace WMDE\Fundraising\Frontend\Infrastructure;
+namespace WMDE\Fundraising\Frontend\Infrastructure\Payment;
 
 /**
  * @license GNU GPL v2+
